refactor: rewrite withTheme decorator by ts
